refactor: add universal details and invocations in superstruct util
